Add `OpaqueError` and `DisplayErrorContext` to `aws-smithy-types`
